requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
librosa
numpy
boto3
pydantic-settings
python-dotenv
